Refactor getLoggedInUserFromLocalStorage
