+++ client/src/api/login/useAutoLogin.ts
@@ -1,22 +1,14 @@
 import { useEffect } from "react";
-import { AuthState } from "../../states/auth";
 import useLogin from "./useLogin";
+import getLoggedInUserFromLocalStorage from "../../util/getLoggedInUserFromLocalStorage";
 
 export default function useAutoLogin() {
   const { login } = useLogin();
   useEffect(() => {
-    try {
-      const savedLoggedInUser = localStorage.getItem("loggedInUser");
-      if (!savedLoggedInUser) {
-        return;
-      }
-      const parsedLoggedInUser: AuthState.LoggedInUser =
-        JSON.parse(savedLoggedInUser);
-      login(parsedLoggedInUser);
-    } catch (e) {
-      console.error("Error parsing loggedInUser from localStorage", e);
-      localStorage.removeItem("loggedInUser");
-      return undefined;
+    const savedLoggedInUser = getLoggedInUserFromLocalStorage();
+    if (!savedLoggedInUser) {
+      return;
     }
+    login(savedLoggedInUser);
   }, []);
 }
